Initial implementation of a basic and overly simplistic HTTP parser.
It currently does a lot of the basic stuff needed but needs some refactoring to make the state machine cleaner. It also needs to be able to handle Transfer-Encoding:chunked, for example, as it currently requires that you give it a Content-Length in a method that should give a body.